refactor: Rename about cloudwatch
