+++ src/spaceone/monitoring/manager/__init__.py
@@ -1,2 +1,2 @@
-from spaceone.monitoring.manager.event_manager import EventManager
+from spaceone.monitoring.manager.cloudwatch_event_manager import EventManager
 from spaceone.monitoring.manager.phd_event_manager import PersonalHealthDashboardManager
