+++ src/spaceone/monitoring/model/cloudwatch_event_response_model.py
